reorganize page_inputs.py as a submodule; move HttpClient to it
