+++ web_poet/page_inputs/__init__.py
@@ -1,0 +1,10 @@
+from .meta import Meta
+from .client import HttpClient
+from .http import (
+    HttpRequest,
+    HttpResponse,
+    HttpRequestHeaders,
+    HttpResponseHeaders,
+    HttpRequestBody,
+    HttpResponseBody,
+)
